requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
pandas>=2.0.0
anthropic>=0.18.0
plotly>=5.15.0
pytest>=7.0.0